Moved new app.py and Dockerfile to challenge folder
