+++ challenge/Dockerfile
@@ -1,12 +1,14 @@
 FROM python:3.9
 
-WORKDIR /usr/src/app
+WORKDIR /app
 
 COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-EXPOSE 5000
+# Change the exposed port to 8080
+EXPOSE 8080
 
+# Update CMD to run the Flask app on port 8080
 CMD ["python", "app.py"]
