+++ challenge/requirements.txt
@@ -1,0 +1,2 @@
+Flask==3.0.3
+redis==5.0.
