+++ requirements.txt
@@ -1,2 +1,1 @@
-Flask==3.0.3
-redis==5.0.
+flask
